Move feature flags synchronization logic out of config package
